+++ lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/graphic/picture/docx_blip/file_reference.rb
@@ -26,6 +26,7 @@ module OoxmlParser
         return self
       end
       return self if @path == 'NULL'
+      return self if @path&.match?(URI::DEFAULT_PARSER.make_regexp)
 
       full_path_to_file = OOXMLDocumentObject.path_to_folder + OOXMLDocumentObject.root_subfolder + @path.gsub('..', '')
       if File.exist?(full_path_to_file)