+++ spec/workbook/worksheet/drawing/picture_spec.rb
@@ -22,4 +22,20 @@ describe 'My behaviour' do
     expect { OoxmlParser::XlsxParser.parse_xlsx('spec/workbook/worksheet/drawing/picture/image_incorrect_link.xlsx') }
       .to output("Cant find path to media file by id: rId1\n").to_stderr
   end
+
+  describe 'file_reference_to_url' do
+    let(:xlsx) { OoxmlParser::XlsxParser.parse_xlsx('spec/workbook/worksheet/drawing/picture/file_reference_to_url.xlsx') }
+    let(:file_reference) do
+      xlsx.worksheets.first.drawings.first
+          .picture.path_to_image.file_reference
+    end
+
+    it 'content of file reference is nil' do
+      expect(file_reference.content).to be_nil
+    end
+
+    it 'path to file reference is url' do
+      expect(file_reference.path).to match(URI::DEFAULT_PARSER.make_regexp)
+    end
+  end
 end